Refactor new array
